+++ packages/yswaps/contracts/interfaces/yearn/IVault.sol
@@ -2,6 +2,8 @@
 pragma solidity >=0.8.4 <0.9.0;
 
 interface IVault {
+  function governance() external view returns (address _governance);
+
   function withdraw(
     uint256,
     address,