+++ packages/yswaps/scripts/libraries/solvers/multicall/IMulticall.ts
@@ -1,0 +1,6 @@
+import { EnabledTrade, TradeSetup } from '@scripts/types';
+
+export interface IMulticall {
+  match(trade: EnabledTrade): boolean;
+  asyncSwap(trade: EnabledTrade): Promise<TradeSetup>;
+}