+++ packages/yswaps/scripts/libraries/solvers/multicall/ThreePoolCrvMulticall.ts
@@ -1,10 +1,10 @@
-import { BigNumber, constants, ethers, PopulatedTransaction, Signer, utils } from 'ethers';
-import { PendingTrade, TradeSetup } from '@scripts/types';
+import { BigNumber, constants, PopulatedTransaction, Signer, utils } from 'ethers';
+import { EnabledTrade, TradeSetup } from '@scripts/types';
 import { IMulticall } from './IMulticall';
 import { ICurveFi, ICurveFi__factory, IERC20, IERC20__factory, IVault, IVault__factory } from '@typechained';
-import { impersonate } from '../utils';
-import zrx from '../libraries/zrx';
-import { mergeTransactions } from '@scripts/libraries/multicall';
+import zrx from '../zrx';
+import { mergeTransactions } from '@scripts/libraries/solvers/multicall';
+import { impersonate } from '@test-utils/wallet';
 
 // 1) 3pool => [usdc|usdt|dai]
 // 2) [usdc|usdt|dai] => yvBOOST
@@ -20,11 +20,11 @@ export class ThreePoolCrvMulticall implements IMulticall {
   private multicallSwapper: string = '0xceB202F25B50e8fAF212dE3CA6C53512C37a01D2';
   private zrxContract: string = '0xDef1C0ded9bec7F1a1670819833240f027b25EfF';
 
-  match(trade: PendingTrade) {
+  match(trade: EnabledTrade) {
     return trade._strategy == this.strategy && trade._tokenIn == this.threeCrv && trade._tokenOut == this.yveCrv;
   }
 
-  async asyncSwap(trade: PendingTrade): Promise<TradeSetup> {
+  async asyncSwap(trade: EnabledTrade): Promise<TradeSetup> {
     const strategySigner: Signer = await impersonate(this.strategy);
     const multicallSwapperSigner: Signer = await impersonate(this.multicallSwapper);
     const crv3Pool: ICurveFi = ICurveFi__factory.connect(this.crv3Pool, multicallSwapperSigner);
@@ -34,13 +34,14 @@ export class ThreePoolCrvMulticall implements IMulticall {
     const yvBoostVault: IVault = IVault__factory.connect(this.yvBoost, multicallSwapperSigner);
     const yveCrvToken: IERC20 = IERC20__factory.connect(this.yveCrv, multicallSwapperSigner);
 
+    const amountIn = '0'; // TODO FIX use strategy balance of in
     console.log('[ThreePoolCrvMulticall] 3crv transfer to swapper');
-    await threeCrv.transfer(this.multicallSwapper, trade._amountIn);
+    await threeCrv.transfer(this.multicallSwapper, amountIn);
 
     // Withdraw usdc from crv3Pool
     console.log('[ThreePoolCrvMulticall] Remove liqudity from curve pool');
     const usdcBalancePre: BigNumber = await usdc.balanceOf(this.multicallSwapper);
-    await crv3Pool.remove_liquidity_one_coin(trade._amountIn, 1, 0);
+    await crv3Pool.remove_liquidity_one_coin(amountIn, 1, 0);
     const usdcBalanceTotal: BigNumber = await usdc.balanceOf(this.multicallSwapper);
     let usdcBalance: BigNumber = usdcBalanceTotal.sub(usdcBalancePre);
     if (usdcBalanceTotal.eq(usdcBalance)) {
@@ -91,7 +92,7 @@ export class ThreePoolCrvMulticall implements IMulticall {
     const transactions: PopulatedTransaction[] = [];
 
     // 1) Withdraw usdc from 3pool
-    transactions.push(await crv3Pool.populateTransaction.remove_liquidity_one_coin(trade._amountIn, 1, 0));
+    transactions.push(await crv3Pool.populateTransaction.remove_liquidity_one_coin(amountIn, 1, 0));
 
     // 2) Approve usdc in zrx (if neccesary)
     if (approveUsdc) transactions.push(await usdc.populateTransaction.approve(zrxAllowanceTarget, constants.MaxUint256));