+++ packages/yswaps/scripts/libraries/solvers/zrx.ts
@@ -82,8 +82,8 @@ export const quote = async (quoteRequest: QuoteRequest): Promise<QuoteResponse> 
       data.minAmountOut = BigNumber.from(`${data.sellAmount}`).mul(rateFromTo).div(BigNumber.from(10).pow(tokenFromDecimals));
     }
   } catch (err: any) {
-    console.log(err.response.data);
-    throw new Error(`Error code: ${err.response.data.code}. Reason: ${err.response.data.reason}`);
+    const code = err.response ? err.response.data.code : err.code;
+    throw new Error(`Error code: ${code}. Reason: ${err.response?.data.reason}`);
   }
   return data;
 };