+++ packages/yswaps/scripts/multicall/IMulticall.ts
@@ -1,6 +1,0 @@
-import { PendingTrade, TradeSetup } from '../types';
-
-export interface IMulticall {
-  match(trade: PendingTrade): boolean;
-  asyncSwap(trade: PendingTrade): Promise<TradeSetup>;
-}