+++ packages/yswaps/scripts/types.ts
@@ -1,5 +1,12 @@
 import { BigNumber } from 'ethers';
 
+export type EnabledTrade = [string, string, string] & {
+  _strategy: string;
+  _tokenIn: string;
+  _tokenOut: string;
+};
+
+// TODO Remove (deprecated)
 export type PendingTrade = [BigNumber, string, string, string, BigNumber, BigNumber] & {
   _id: BigNumber;
   _strategy: string;